move test to other module
